refactor: extract svg and mermaid preview into shared components
